[FEATURE] Standalone Version
Adding files as arguments ,optional file export, proper requirements.txt
