+++ README.md
@@ -1,24 +1,37 @@
 # Decrypt Chrome Passwords
-A simple program to decrypt chrome password saved on your machine. <br>
-This code has only been tested on windows, so it may not work on other OS.<br>
-If you have an idea for improvement, do let me know!<br>
+
+Decrypt chrome password from a given "Login Data" chrome database and the "Local State" JSON file that contains the key and optionally export output to a CSV file.
+
+## Installation
+```
+git clone https://github.com/Bruckyy/chrome-password-decrypter.git
+cd chrome-password-decrypter
+pip install -r requirements.txt
+```
+
+## Usage
+```
+> python .\chrome-decrypt.py --help
+usage: chrome-decrypt.py [-h] -k KEY -d DB [-o OUTPUT]
+
+> python chrome-decrypt.py --key local_state_file --db login_data_file -o decrypted_passwords
+
+**************************************************
+Sequence: 1
+URL: https://www.github.com/login
+User Name: <USERNAME>
+Password: <PASSWORD>
+
+**************************************************
+<SNIP>
+...
+
+Exported decrypted passwords to decrypted_passwords.csv
+
+```
 
 ## OS support
 1. Windows
 
-## Dependencies (see requirements)
-1. sqlite
-2. pycryptodomex
-3. pywin32
-
-## Usage
-python decrypt_chrome_password.py<br>
-
-## Output
-Saved as decrypted_password.csv
-
-## Medium
-To understand the how this program works, read my medium article. <br>
-https://ohyicong.medium.com/how-to-hack-chrome-password-with-python-1bedc167be3d
-
-
+## Details
+This project is a standalone version that takes the files as arguments and a fork of [this](https://github.com/ohyicong/decrypt-chrome-passwords) project
